requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
imbalanced-learn==0.11.0

# Streamlit dashboard
streamlit==1.28.1
plotly==5.17.0

# Visualization (for notebook)
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
